# api/swagger.yaml
basePath: /
definitions:
...
  httpx.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  httpx.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.APIError'
    type: object
  menu.CreateMenuRecipeRequest:
    properties:
      authorEmployeeId:
        type: string
      badge:
        example: ХИТ
        type: string
      description:
        type: string
      name:
        example: Negroni
        type: string
      price:
        example: 550
        type: number
      rating:
        example: 4.6
        type: number
      restaurantId:
        type: string
      strength:
        example: 7
        type: integer
      tags:
        items:
          type: string
        type: array
    type: object
  menu.ListMenuRequest:
    properties:
      restaurantId:
        type: string
    type: object
  menu.MenuRecipeInput:
    properties:
      badge:
        example: ХИТ
        type: string
      description:
        type: string
      name:
        example: Negroni
        type: string
      price:
        example: 550
        type: number
      rating:
        example: 4.6
        type: number
      strength:
        example: 7
        type: integer
      tags:
        items:
          type: string
        type: array
    type: object
  menu.MenuRecipeView:
    properties:
      authorEmployeeId:
        type: string
      badge:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      rating:
        type: number
      restaurantId:
        type: string
      strength:
        type: integer
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: Backend for the mixMaster restaurant ordering / recipes app.
  title: mixMaster API
  version: 0.1.0
paths:
...
  /menu:
    post:
      consumes:
      - application/json
      parameters:
      - description: Menu recipe
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/menu.CreateMenuRecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/menu.MenuRecipeView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Create a menu recipe
      tags:
      - menu
  /menu/{id}:
    delete:
      parameters:
      - description: Menu recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Remove a menu recipe (soft)
      tags:
      - menu
    patch:
      consumes:
      - application/json
      parameters:
      - description: Menu recipe ID
        in: path
        name: id
        required: true
        type: string
      - description: New values
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/menu.MenuRecipeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/menu.MenuRecipeView'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Edit a menu recipe (full replace of editable fields)
      tags:
      - menu
  /menu/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Restaurant id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/menu.ListMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/menu.MenuRecipeView'
                  type: array
              type: object
      summary: List a restaurant's menu recipes
      tags:
      - menu
...
swagger: "2.0"